+++ spring-app/src/main/java/org/example/tmsserver/service/RegionService.java
@@ -4,6 +4,8 @@ import org.example.tmsserver.entity.Region;
 import org.example.tmsserver.repository.RegionRepository;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 public class RegionService {
 
@@ -13,7 +15,7 @@ public class RegionService {
         this.regionRepository = regionRepository;
     }
 
-    public Region getRegionByPoint(double lat, double lon) {
+    public Region getRegionByPoint(BigDecimal lat, BigDecimal lon) {
         Integer regionId = regionRepository.findRegionByPoint(lat, lon);
         if (regionId != null) {
             return regionRepository.findById(regionId).orElse(null);
